vim: Unsubmodule-ize the plugins
I don't use vim as my main editor anymore. I don't really care if the
plugins are up to date. I'd rather just see that cloning this repository
is all that's needed to get a passable experience in vim. This means
that I don't want to download a large amount of plugins for every new
computer I put these dotfiles on.